refactor(voice): rewrite imports sdk/speech/* and sdkx/{stt,tts}/* -> voice/*
Mechanical, single-pass perl rewrite across voice/ and examples/:

  github.com/GizClaw/flowcraft/sdk/speech         -> .../voice
  github.com/GizClaw/flowcraft/sdkx/stt/bytedance -> .../voice/stt/bytedance
  github.com/GizClaw/flowcraft/sdkx/tts/minimax   -> .../voice/tts/minimax

43 files touched, only import paths changed. Identifier-level usage of
the speech.* root package symbols (NewPipeline, STT, TTS, etc.) is
addressed in the next commit, which also bumps the package declaration
from `package speech` to `package voice`.

The voice/go.mod still pins sdk v0.1.12 published; go.work resolves
all flowcraft.* paths to local source, so the tree begins to compile
again from this commit onward (modulo the speech -> voice symbol
rename, handled next).

Made-with: Cursor
