+++ voice/pipeline.go
@@ -9,10 +9,10 @@ import (
 	"time"
 
 	"github.com/GizClaw/flowcraft/sdk/model"
-	"github.com/GizClaw/flowcraft/sdk/speech/audio"
-	"github.com/GizClaw/flowcraft/sdk/speech/provider"
-	"github.com/GizClaw/flowcraft/sdk/speech/stt"
-	"github.com/GizClaw/flowcraft/sdk/speech/tts"
+	"github.com/GizClaw/flowcraft/voice/audio"
+	"github.com/GizClaw/flowcraft/voice/provider"
+	"github.com/GizClaw/flowcraft/voice/stt"
+	"github.com/GizClaw/flowcraft/voice/tts"
 	"github.com/GizClaw/flowcraft/sdk/workflow"
 	"github.com/rs/xid"
 )
