+++ voice/pipeline_test.go
@@ -10,11 +10,11 @@ import (
 	"testing"
 	"time"
 
-	"github.com/GizClaw/flowcraft/sdk/speech"
-	"github.com/GizClaw/flowcraft/sdk/speech/audio"
-	"github.com/GizClaw/flowcraft/sdk/speech/provider"
-	"github.com/GizClaw/flowcraft/sdk/speech/stt"
-	"github.com/GizClaw/flowcraft/sdk/speech/tts"
+	"github.com/GizClaw/flowcraft/voice"
+	"github.com/GizClaw/flowcraft/voice/audio"
+	"github.com/GizClaw/flowcraft/voice/provider"
+	"github.com/GizClaw/flowcraft/voice/stt"
+	"github.com/GizClaw/flowcraft/voice/tts"
 	"github.com/GizClaw/flowcraft/sdk/workflow"
 )
