+++ voice/resample_pipeline_test.go
@@ -5,8 +5,8 @@ import (
 	"math"
 	"testing"
 
-	"github.com/GizClaw/flowcraft/sdk/speech/audio"
-	"github.com/GizClaw/flowcraft/sdk/speech/stt"
+	"github.com/GizClaw/flowcraft/voice/audio"
+	"github.com/GizClaw/flowcraft/voice/stt"
 )
 
 func makePCM16SineResampleTest(freq float64, sampleRate, numSamples, channels int) []byte {
