+++ voice/session_test.go
@@ -10,13 +10,13 @@ import (
 	"testing"
 	"time"
 
-	"github.com/GizClaw/flowcraft/sdk/speech"
-	"github.com/GizClaw/flowcraft/sdk/speech/audio"
-	"github.com/GizClaw/flowcraft/sdk/speech/detect"
-	speechmetrics "github.com/GizClaw/flowcraft/sdk/speech/metrics"
-	"github.com/GizClaw/flowcraft/sdk/speech/preprocess"
-	"github.com/GizClaw/flowcraft/sdk/speech/stt"
-	"github.com/GizClaw/flowcraft/sdk/speech/tts"
+	"github.com/GizClaw/flowcraft/voice"
+	"github.com/GizClaw/flowcraft/voice/audio"
+	"github.com/GizClaw/flowcraft/voice/detect"
+	speechmetrics "github.com/GizClaw/flowcraft/voice/metrics"
+	"github.com/GizClaw/flowcraft/voice/preprocess"
+	"github.com/GizClaw/flowcraft/voice/stt"
+	"github.com/GizClaw/flowcraft/voice/tts"
 	"github.com/GizClaw/flowcraft/sdk/workflow"
 )
