+++ voice/tts/fallback.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"io"
 
-	"github.com/GizClaw/flowcraft/sdk/speech/audio"
-	"github.com/GizClaw/flowcraft/sdk/speech/provider"
+	"github.com/GizClaw/flowcraft/voice/audio"
+	"github.com/GizClaw/flowcraft/voice/provider"
 )
 
 // FallbackTTS tries providers in order until one succeeds.
