+++ voice/webrtc/sink.go
@@ -5,8 +5,8 @@ import (
 	"io"
 	"time"
 
-	"github.com/GizClaw/flowcraft/sdk/speech/audio"
-	"github.com/GizClaw/flowcraft/sdk/speech/tts"
+	"github.com/GizClaw/flowcraft/voice/audio"
+	"github.com/GizClaw/flowcraft/voice/tts"
 	"github.com/GizClaw/flowcraft/sdk/telemetry"
 	"github.com/pion/webrtc/v4/pkg/media"
 	"go.opentelemetry.io/otel/metric"
